Add custom errors with metadata on ChatConnector
This enables the cleanup of potencial invalid data.

Use case:

When a channel is deleted in Slack there is no event that gets triggered
into the bot. This means that any address associated with the channel
that is cached no longer will be valid. When trying to send a message to
this address, the global handler error will emit an error without
metadata about the action that triggered the error. Using this new type of error, the metadata can
be introspected and be used for taking actions.