Add startReplyChain to UniversalBot and IConnector
At the moment the following connectors implement the startReplyChain:

- botbuilder-teams
  (https://github.com/OfficeDev/BotBuilder-MicrosoftTeams)
- botbuilder-slack (https://github.com/suttna/botbuilder-slack)

On both cases, the current solution at the connector level breaks the
usage of middlewares. In both cases, the "send" middleware is not called
when sending a message using the startReplyChain api.

We want to standarize the startReplyChain since a lot of clients support
it and is very useful operation. The idea would be to get this merged so
that both previously mentioned libraries can update the implementation
to match the new signature.